split curricula tests into multiple files
